try replacing symlinks
